feat(base64)!: Update packaging, expose encode/decode entry modules, bridge RESM/NESM
BREAKING CHANGE: No longer supports direct use from CommonJS
