+++ packages/base64/decode.js
@@ -1,0 +1,1 @@
+export { decodeBase64 } from './src/decode.js';
