+++ packages/base64/encode.js
@@ -1,0 +1,1 @@
+export { encodeBase64 } from './src/encode.js';
