+++ packages/base64/index.d.ts
@@ -1,0 +1,1 @@
+export { encodeBase64, decodeBase64 } from './index.js';
