+++ packages/base64/index.js
@@ -1,0 +1,2 @@
+export { encodeBase64 } from './src/encode.js';
+export { decodeBase64 } from './src/decode.js';
