+++ packages/base64/src/decode.js
@@ -1,0 +1,55 @@
+// @ts-check
+/* eslint no-bitwise: ["off"] */
+
+import { monodu64, padding } from './common.js';
+
+/**
+ * Decodes a Base64 string into bytes, as specified in
+ * https://tools.ietf.org/html/rfc4648#section-4
+ *
+ * @param {string} string Base64-encoded string
+ * @param {string} [name] The name of the string as it will appear in error
+ * messages.
+ * @returns {Uint8Array} decoded bytes
+ */
+export function decodeBase64(string, name = '<unknowwn>') {
+  const data = new Uint8Array(Math.ceil((string.length * 4) / 3));
+  let register = 0;
+  let quantum = 0;
+  let i = 0; // index in string
+  let j = 0; // index in data
+
+  while (i < string.length && string[i] !== padding) {
+    const number = monodu64[string[i]];
+    if (number === undefined) {
+      throw Error(`Invalid base64 character ${string[i]} in string ${name}`);
+    }
+    register = (register << 6) | number;
+    quantum += 6;
+    if (quantum >= 8) {
+      quantum -= 8;
+      data[j] = register >>> quantum;
+      j += 1;
+      register &= (1 << quantum) - 1;
+    }
+    i += 1;
+  }
+
+  while (i < string.length && quantum % 8 !== 0) {
+    if (string[i] !== padding) {
+      throw Error(`Missing padding at offset ${i} of string ${name}`);
+    }
+    i += 1;
+    quantum += 6;
+  }
+
+  if (i < string.length) {
+    throw Error(
+      `Base64 string has trailing garbage ${string.substr(
+        i,
+      )} in string ${name}`,
+    );
+  }
+
+  return data.subarray(0, j);
+}
